#72 Fix code complexity in Psalm Drupal autoloader.
